+++ spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/freemarker/FreeMarkerAutoConfiguration.java
@@ -18,36 +18,24 @@ package org.springframework.boot.autoconfigure.freemarker;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Properties;
 
 import javax.annotation.PostConstruct;
-import javax.servlet.Servlet;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnNotWebApplication;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication.Type;
 import org.springframework.boot.autoconfigure.template.TemplateLocation;
-import org.springframework.boot.autoconfigure.web.ConditionalOnEnabledResourceChain;
-import org.springframework.boot.autoconfigure.web.servlet.WebMvcAutoConfiguration;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.ui.freemarker.FreeMarkerConfigurationFactory;
 import org.springframework.ui.freemarker.FreeMarkerConfigurationFactoryBean;
-import org.springframework.web.servlet.resource.ResourceUrlEncodingFilter;
-import org.springframework.web.servlet.view.freemarker.FreeMarkerConfig;
-import org.springframework.web.servlet.view.freemarker.FreeMarkerConfigurer;
-import org.springframework.web.servlet.view.freemarker.FreeMarkerViewResolver;
 
 /**
  * {@link EnableAutoConfiguration Auto-configuration} for FreeMarker.
@@ -58,10 +46,10 @@ import org.springframework.web.servlet.view.freemarker.FreeMarkerViewResolver;
  * @since 1.1.0
  */
 @Configuration
-@ConditionalOnClass({ freemarker.template.Configuration.class,
-		FreeMarkerConfigurationFactory.class })
-@AutoConfigureAfter(WebMvcAutoConfiguration.class)
+@ConditionalOnClass({freemarker.template.Configuration.class,
+		FreeMarkerConfigurationFactory.class})
 @EnableConfigurationProperties(FreeMarkerProperties.class)
+@Import({FreeMarkerServletWebConfiguration.class, FreeMarkerReactiveWebConfiguration.class})
 public class FreeMarkerAutoConfiguration {
 
 	private static final Log logger = LogFactory
@@ -99,25 +87,13 @@ public class FreeMarkerAutoConfiguration {
 		}
 	}
 
-	protected static class FreeMarkerConfiguration {
-
-		@Autowired
-		protected FreeMarkerProperties properties;
-
-		protected void applyProperties(FreeMarkerConfigurationFactory factory) {
-			factory.setTemplateLoaderPaths(this.properties.getTemplateLoaderPath());
-			factory.setPreferFileSystemAccess(this.properties.isPreferFileSystemAccess());
-			factory.setDefaultEncoding(this.properties.getCharsetName());
-			Properties settings = new Properties();
-			settings.putAll(this.properties.getSettings());
-			factory.setFreemarkerSettings(settings);
-		}
-
-	}
-
 	@Configuration
 	@ConditionalOnNotWebApplication
-	public static class FreeMarkerNonWebConfiguration extends FreeMarkerConfiguration {
+	public static class FreeMarkerNonWebConfiguration extends AbstractFreeMarkerConfiguration {
+
+		public FreeMarkerNonWebConfiguration(FreeMarkerProperties properties) {
+			super(properties);
+		}
 
 		@Bean
 		@ConditionalOnMissingBean
@@ -125,43 +101,6 @@ public class FreeMarkerAutoConfiguration {
 			FreeMarkerConfigurationFactoryBean freeMarkerFactoryBean = new FreeMarkerConfigurationFactoryBean();
 			applyProperties(freeMarkerFactoryBean);
 			return freeMarkerFactoryBean;
-		}
-
-	}
-
-	@Configuration
-	@ConditionalOnClass({ Servlet.class, FreeMarkerConfigurer.class })
-	@ConditionalOnWebApplication(type = Type.SERVLET)
-	public static class FreeMarkerWebConfiguration extends FreeMarkerConfiguration {
-
-		@Bean
-		@ConditionalOnMissingBean(FreeMarkerConfig.class)
-		public FreeMarkerConfigurer freeMarkerConfigurer() {
-			FreeMarkerConfigurer configurer = new FreeMarkerConfigurer();
-			applyProperties(configurer);
-			return configurer;
-		}
-
-		@Bean
-		public freemarker.template.Configuration freeMarkerConfiguration(
-				FreeMarkerConfig configurer) {
-			return configurer.getConfiguration();
-		}
-
-		@Bean
-		@ConditionalOnMissingBean(name = "freeMarkerViewResolver")
-		@ConditionalOnProperty(name = "spring.freemarker.enabled", matchIfMissing = true)
-		public FreeMarkerViewResolver freeMarkerViewResolver() {
-			FreeMarkerViewResolver resolver = new FreeMarkerViewResolver();
-			this.properties.applyToViewResolver(resolver);
-			return resolver;
-		}
-
-		@Bean
-		@ConditionalOnMissingBean
-		@ConditionalOnEnabledResourceChain
-		public ResourceUrlEncodingFilter resourceUrlEncodingFilter() {
-			return new ResourceUrlEncodingFilter();
 		}
 
 	}
